+++ README.md
@@ -14,16 +14,15 @@
 - [x] 자동차는 전진 or 멈출 수 있다.
 
 ### 제약 조건
-- [ ] 자동차의 이름은 1자 이상 5자 이하여야 한다.
-- [ ] 자동차 이름은 쉼표(,)를 기준으로 구분하여 입력받는다.
-- [ ] 게임을 진행할 횟수는 0 이상이어야 한다.
+- [x] 자동차의 이름은 1자 이상 5자 이하여야 한다.
+- [x] 자동차 이름은 쉼표(,)를 기준으로 구분하여 입력받는다.
 
 ### 입출력
-- [ ] 자동차 이름을 입력받는다. (쉼표를 기준으로)
-- [ ] 게임을 진행할 횟수를 입력받는다.
-- [ ] 실행 결과에는 자동차 이름과 위치(-)를 표시한다.
-- [ ] 게임이 완료되면 누가 우승했는지를 출력한다.
+- [x] 자동차 이름을 입력받는다. (쉼표를 기준으로)
+- [x] 게임을 진행할 횟수를 입력받는다.
+- [x] 실행 결과에는 자동차 이름과 위치(-)를 표시한다.
+- [x] 게임이 완료되면 누가 우승했는지를 출력한다.
 
 ### 게임 진행
-- [ ] 0과 9 사이에서 랜덤값을 구한 후 랜덤값이 4 이상일 경우 전진하고, 3 이하의 값이면 멈춘다.
-- [ ] 위와 같은 프로세스를 입력 받은 게임 횟수만큼 진행한다.
+- [x] 0과 9 사이에서 랜덤값을 구한 후 랜덤값이 4 이상일 경우 전진하고, 3 이하의 값이면 멈춘다.
+- [x] 위와 같은 프로세스를 입력 받은 게임 횟수만큼 진행한다.
