+++ src/main/java/CarName.java
@@ -7,6 +7,7 @@ public class CarName {
     private static final int MIN_CAR_NAME_LENGTH = 1;
     private static final int MAX_CAR_NAME_LENGTH = 5;
     private static final String NAME_SPLITTER = ",";
+
     private String name;
 
     public CarName(String name) {
@@ -19,7 +20,10 @@ public class CarName {
 
     // 자동차 이름 길이 검증 메서드
     public static boolean validateNameLength(String name) {
-        return name.length() >= MIN_CAR_NAME_LENGTH && name.length() <= MAX_CAR_NAME_LENGTH;
+        if (name.length() >= MIN_CAR_NAME_LENGTH && name.length() <= MAX_CAR_NAME_LENGTH) {
+            return true;
+        }
+        throw new IllegalArgumentException("자동차 이름의 길이가 올바르지 않습니다.");
     }
 
     public String getCarName() {
