+++ src/main/java/NumGenerator.java
@@ -3,6 +3,7 @@ import java.util.Random;
 public class NumGenerator {
 
     private static final int NUMBER_MAX_VALUE = 10;
+
     private int number;
 
     Random random = new Random();
